Re-organizing project
Fix next/link stuffs

Isolate styles from components and pages

Isolate styles and minor fix proj structure

Refactoring pain

Rename folders
