+++ src/components/Carousels/ArtCarousel.module.scss
@@ -1,0 +1,21 @@
+@use "partials/mixins" as *;
+
+#carousel-comp {
+  padding-top: 25px;
+  #carousel-container {
+    display: flex;
+    padding: 25px 0 100px 0;
+    flex-direction: row;
+    justify-content: left;
+    h2 {
+      font-size: 2rem;
+    }
+    #carousel {
+      display: flex;
+      flex-wrap: wrap;
+      flex-direction: row;
+      // Two rows per page
+      overflow-x: hidden;
+    }
+  }
+}
