+++ src/components/Carousels/Carousel.module.scss
