+++ src/components/Gallery/Artwork.module.scss
