+++ src/components/Gallery/GalleryCarousel.module.scss
