+++ src/components/Hero/Hero.module.scss
@@ -1,6 +1,6 @@
 @use "@/styles/partials/mixins" as *;
 
-/* Cards animation by skepfusky */
+/* Cards animation by Kuroji Fusky */
 .hero-blur-gradient-overlay {
   $blurAmount: 10rem;
   position: absolute;
