+++ src/components/Hero/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./Hero"
