--- src/components/layouts/Container.tsx
@@ -1,9 +1,0 @@
-import { ReactNode } from "react"
-
-interface ContainerProps {
-  children: ReactNode
-}
-
-export default function Container({ children }: ContainerProps) {
-  return <div className="container">{children}</div>
-}
