+++ src/pages/index.tsx
@@ -1,9 +1,8 @@
 import React, { useEffect, useState } from "react"
-import Container from "@/components/layouts/Container"
+import Container from "@/components/Layouts/Container"
 import styles from "@/styles/Home.module.scss"
-import Carousel from "@/components/carousel/CarouselMenu"
-import MetaTags from "@/components/layouts/MetaTags"
-import StackedCards from "../components/home/StackedCards"
+import Carousel from "@/components/Carousels/CarouselMenu"
+import StackedCards from "../components/Hero/Hero"
 
 export default function Home() {
   const [data, setData] = useState<Fursona[]>([])
@@ -30,27 +29,24 @@ export default function Home() {
   }, [])
   return (
     <Container>
-      <MetaTags />
-      <main>
-        <div id={styles["header-hero"]}>
-          <article>
-            <h1>Manage, store, and show your fursonas in one place</h1>
-            <p>
-              <strong>MyFursona</strong> is a platform where you can show off
-              your fluffy characters and show them off to your friends!
-            </p>
-          </article>
-          <StackedCards />
-        </div>
-        <div className={styles["fursona-main-showcase"]}>
-          <section aria-label="Popular Fursonas">
-            <Carousel title="Popular Fursonas" type="popular" items={data} />
-          </section>
-          <section aria-label="New Fursonas" id={styles["card-previews"]}>
-            <Carousel title={"New Fursonas"} type="new" />
-          </section>
-        </div>
-      </main>
+      <div id={styles["header-hero"]}>
+        <article>
+          <h1>Manage, store, and show your fursonas in one place</h1>
+          <p>
+            <strong>MyFursona</strong> is a platform where you can show off your
+            fluffy characters and show them off to your friends!
+          </p>
+        </article>
+        <StackedCards />
+      </div>
+      <div className={styles["fursona-main-showcase"]}>
+        <section aria-label="Popular Fursonas">
+          <Carousel title="Popular Fursonas" type="popular" items={data} />
+        </section>
+        <section aria-label="New Fursonas" id={styles["card-previews"]}>
+          <Carousel title={"New Fursonas"} type="new" />
+        </section>
+      </div>
     </Container>
   )
 }
