+++ src/pages/profile/Profile.module.scss
@@ -1,4 +1,4 @@
-@use "partials/mixins" as *;
+@use "@/styles/partials/mixins" as *;
 
 #user-container-loading {
   display: flex;
