+++ src/pages/profile/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./profile"
